GCS/requirements.txt
# GCS Component Requirements
pymavlink>=2.4.37
pyyaml>=6.0
